completed create petition activity

+++ app/src/main/java/com/gateszeng/sdhacks2016/Petition.java
@@ -51,7 +51,7 @@ public class Petition {
         this.votes = votes;
     }
 
-    public int getTime() {
+    public long getTime() {
         return time;
     }
